+++ docs/Whats-New.md
@@ -403,6 +403,8 @@ Vanilla fixes:
 - Fixed a glitch related to incorrect target setting for missiles (by Belonit)
 - Skipped parsing `[Header]` section of compaign maps which led to occasional crashes on Linux (by Trsdy)
 - Fixed units' turret rotation and jumpjet wobble under EMP (by Trsdy)
+- Fixed weapon effects like railguns, lasers, beams, bolts and waves drawing beyond where Warhead detonates if it hits an obstacle like wall (by Starkku)
+- Fixed `AmbientDamage` when used with `IsRailgun=yes` being cut off by elevation changes (by Starkku)
 
 Phobos fixes:
 - Fixed a few errors of calling for superweapon launch by `LaunchSW` or building infiltration (by Trsdy)